_external_review/learn_zip/requirements.txt
streamlit
pandas
sqlalchemy
requests
pyspark
python-dotenv
